refactor: api instance 이름 명확화 및 불필요 코드 삭제
